# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Comercial:
    properties:
      categoria:
        type: string
      id:
        type: integer
      marca:
        type: string
      nome:
        type: string
      preco_venda:
        type: number
    type: object
  model.ComercialCreate:
    properties:
      categoria:
        type: string
      marca:
        type: string
      nome:
        type: string
      preco_venda:
        type: number
    type: object
...
  model.Produto:
    properties:
      categoria:
        type: string
      id:
        type: integer
      marca:
        type: string
      nome:
        type: string
    type: object
  model.ProdutoCreate:
    properties:
      categoria:
        type: string
      marca:
        type: string
      nome:
        type: string
    type: object
  model.ProdutoWithQnt:
    properties:
      categoria:
        type: string
      id:
        type: integer
      marca:
        type: string
      nome:
        type: string
      quantidade_disponível:
        type: integer
    type: object
...
  model.UnionProduto:
    properties:
      categoria:
        type: string
      id:
        type: integer
      marca:
        type: string
      nome:
        type: string
      preco_venda:
        type: number
    type: object
...
  types.ErrorResponse:
    properties:
      detail:
        type: string
    type: object
info:
  contact: {}
  description: Aplicação de Banco de Dados para Gerenciamento de Bares
  title: EDNA Bar System
  version: "1.0"
paths:
...
  /produtos:
    get:
      parameters:
      - description: 'Filter by nome. Format: <op>.<value>. Ops: like, ilike, eq,
          ne'
        in: query
        name: filter-nome
        type: string
      - description: 'Filter by categoria. Format: <op>.<value>. Ops: like, ilike,
          eq, ne'
        in: query
        name: filter-categoria
        type: string
      - description: 'Filter by marca. Format: <op>.<value>. Ops: like, ilike, eq,
          ne'
        in: query
        name: filter-marca
        type: string
      - description: 'Sort by attribute. Allowed: nome, categoria, marca. Prefix ''-''
          for desc. Comma separated'
        in: query
        name: sort
        type: string
      - description: Pagination offset (default 0)
        in: query
        name: offset
        type: integer
      - description: Pagination limit (default 0)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.UnionProduto'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: List Produtos (all types)
      tags:
      - Produtos
    post:
      consumes:
      - application/json
      parameters:
      - description: Product payload
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/model.ProdutoCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Produto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Create Produto
      tags:
      - Produtos
  /produtos/{id}:
    delete:
      parameters:
      - description: Produto ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Delete Produto
      tags:
      - Produtos
    get:
      parameters:
      - description: Produto ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Produto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get Produto by ID
      tags:
      - Produtos
    put:
      consumes:
      - application/json
      parameters:
      - description: Produto ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product payload
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/model.ProdutoCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Produto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Update Produto
      tags:
      - Produtos
  /produtos/{id}/quantidades:
    get:
      parameters:
      - description: Produto ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ProdutoWithQnt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get Produto Quantidade
      tags:
      - Produtos
  /produtos/comercial:
    get:
      parameters:
      - description: 'Filter by nome. Format: <op>.<value>. Ops: like, ilike, eq,
          ne'
        in: query
        name: filter-nome
        type: string
      - description: 'Filter by categoria. Format: <op>.<value>. Ops: like, ilike,
          eq, ne'
        in: query
        name: filter-categoria
        type: string
      - description: 'Filter by marca. Format: <op>.<value>. Ops: like, ilike, eq,
          ne'
        in: query
        name: filter-marca
        type: string
      - description: 'Filter by preco_venda. Format: <op>.<value>. Ops: eq, ne, lt,
          gt, le, ge'
        in: query
        name: filter-preco_venda
        type: number
      - description: 'Sort fields: nome, categoria, marca, preco_venda. Prefix ''-''
          for desc. Comma separated'
        in: query
        name: sort
        type: string
      - description: Pagination offset (default 0)
        in: query
        name: offset
        type: integer
      - description: Pagination limit (default 0)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Comercial'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: List Comercial products
      tags:
      - Produtos
    post:
      consumes:
      - application/json
      parameters:
      - description: Comercial product payload
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/model.ComercialCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Comercial'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Create Comercial Produto
      tags:
      - Produtos
  /produtos/comercial/{id}:
    get:
      parameters:
      - description: Produto ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Comercial'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get Comercial Produto by ID
      tags:
      - Produtos
    put:
      consumes:
      - application/json
      parameters:
      - description: Produto ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comercial product payload
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/model.ComercialCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Comercial'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Update Comercial Produto
      tags:
      - Produtos
  /produtos/estrutural:
    get:
      parameters:
      - description: 'Filter by nome. Format: <op>.<value>. Ops: like, ilike, eq,
          ne'
        in: query
        name: filter-nome
        type: string
      - description: 'Filter by categoria. Format: <op>.<value>. Ops: like, ilike,
          eq, ne'
        in: query
        name: filter-categoria
        type: string
      - description: 'Filter by marca. Format: <op>.<value>. Ops: like, ilike, eq,
          ne'
        in: query
        name: filter-marca
        type: string
      - description: 'Sort fields: nome, categoria, marca. Prefix ''-'' for desc.
          Comma separated'
        in: query
        name: sort
        type: string
      - description: Pagination offset (default 0)
        in: query
        name: offset
        type: integer
      - description: Pagination limit (default 0)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Produto'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: List Estrutural products
      tags:
      - Produtos
...
swagger: "2.0"